preprocessor/requirements.txt
PIKA==1.2.0
PYSMB==1.2.6
PILLOW==8.1.0
PYTHON-DOTENV==0.15.0